69457 - Updating Org Model so that UKPRN is an int.

+++ src/ESFA.DC.ILR.ReferenceDataService.Data.Population/Interface/IOrganisationsDataRetrievalService.cs
@@ -7,6 +7,6 @@ namespace ESFA.DC.ILR.ReferenceDataService.Data.Population.Interface
 {
     public interface IOrganisationsDataRetrievalService
     {
-        Task<IReadOnlyDictionary<long, Organisation>> RetrieveAsync(IReadOnlyCollection<int> ukprns, CancellationToken cancellationToken);
+        Task<IReadOnlyDictionary<int, Organisation>> RetrieveAsync(IReadOnlyCollection<int> ukprns, CancellationToken cancellationToken);
     }
 }
